first feature, adding pets
